requirements.txt
django
psycopg2
python-dotenv
requests
djangorestframework